+++ NEWS.md
@@ -1,3 +1,20 @@
+# table1 1.4
+
+* Headline feature: ability to convert `table1` output to a `data.frame`,
+  `flextable` or `kableExtra` object (though with some limitations). As a
+  direct consequence, `table1` can now be used within LaTeX documents, and
+  there is a new vignette to document this feature. Output to .docx is also
+  possible with this feature (through pandoc or `flextable`).
+
+* Fomulas containing dot ('.') are now handled correctly (the '.' means "all
+  variables other than those that appear elsewhere in the formula").
+
+* Arguments to formatC (e.g. big.mark) will apply to all numbers by default
+  (previously, only continuous variables had this).
+
+* New `setLabel()` function provides an alternative syntax for setting labels
+  (useful for setting a label inside a formula).
+
 # table1 1.3
 
 * The headline feature is the `extra.col` option, which gives a flexible way to